change deprecated tf.to_float to tf.cast

+++ models-r1.13.0/official/transformer/transformer_main.py
@@ -146,8 +146,10 @@ def record_scalars(metric_dict):
 def get_learning_rate(learning_rate, hidden_size, learning_rate_warmup_steps):
   """Calculate learning rate with linear warmup and rsqrt decay."""
   with tf.name_scope("learning_rate"):
-    warmup_steps = tf.to_float(learning_rate_warmup_steps)
-    step = tf.to_float(tf.train.get_or_create_global_step())
+    # warmup_steps = tf.to_float(learning_rate_warmup_steps)
+    warmup_steps = tf.cast(learning_rate_warmup_steps, tf.float32)
+    # step = tf.to_float(tf.train.get_or_create_global_step())
+    step = tf.cast(tf.train.get_or_create_global_step(), tf.float32)
 
     learning_rate *= (hidden_size ** -0.5)
     # Apply linear warmup
